+++ applications/main/subghz/scenes/subghz_scene_rpc.c
@@ -3,6 +3,7 @@
 #include <lib/subghz/protocols/star_line.h>
 #include <lib/subghz/protocols/alutech_at_4n.h>
 #include <lib/subghz/protocols/nice_flor_s.h>
+#include <lib/subghz/protocols/somfy_telis.h>
 
 typedef enum {
     SubGhzRpcStateIdle,
@@ -115,6 +116,7 @@ void subghz_scene_rpc_on_exit(void* context) {
     keeloq_reset_original_btn();
     alutech_reset_original_btn();
     nice_flors_reset_original_btn();
+    somfy_telis_reset_original_btn();
     star_line_reset_mfname();
     star_line_reset_kl_type();
 }