+++ applications/main/subghz/views/transmitter.c
@@ -7,6 +7,7 @@
 #include <lib/subghz/protocols/keeloq.h>
 #include <lib/subghz/protocols/alutech_at_4n.h>
 #include <lib/subghz/protocols/nice_flor_s.h>
+#include <lib/subghz/protocols/somfy_telis.h>
 
 struct SubGhzViewTransmitter {
     View* view;
@@ -161,6 +162,7 @@ bool subghz_view_transmitter_input(InputEvent* event, void* context) {
         keeloq_set_btn(1);
         alutech_set_btn(1);
         nice_flors_set_btn(1);
+        somfy_telis_set_btn(1);
         with_view_model(
             subghz_transmitter->view,
             SubGhzViewTransmitterModel * model,
@@ -175,6 +177,10 @@ bool subghz_view_transmitter_input(InputEvent* event, void* context) {
                 } else if(nice_flors_get_original_btn() != 0) {
                     furi_string_printf(
                         model->temp_button_id, "%01X", nice_flors_get_original_btn());
+                    model->draw_temp_button = true;
+                } else if(somfy_telis_get_original_btn() != 0) {
+                    furi_string_printf(
+                        model->temp_button_id, "%01X", somfy_telis_get_original_btn());
                     model->draw_temp_button = true;
                 }
             },
@@ -192,6 +198,7 @@ bool subghz_view_transmitter_input(InputEvent* event, void* context) {
         keeloq_set_btn(2);
         alutech_set_btn(2);
         nice_flors_set_btn(2);
+        somfy_telis_set_btn(2);
         with_view_model(
             subghz_transmitter->view,
             SubGhzViewTransmitterModel * model,
@@ -206,6 +213,10 @@ bool subghz_view_transmitter_input(InputEvent* event, void* context) {
                 } else if(nice_flors_get_original_btn() != 0) {
                     furi_string_printf(
                         model->temp_button_id, "%01X", nice_flors_get_original_btn());
+                    model->draw_temp_button = true;
+                } else if(somfy_telis_get_original_btn() != 0) {
+                    furi_string_printf(
+                        model->temp_button_id, "%01X", somfy_telis_get_original_btn());
                     model->draw_temp_button = true;
                 }
             },
@@ -223,6 +234,7 @@ bool subghz_view_transmitter_input(InputEvent* event, void* context) {
         keeloq_set_btn(3);
         alutech_set_btn(3);
         nice_flors_set_btn(3);
+        somfy_telis_set_btn(3);
         with_view_model(
             subghz_transmitter->view,
             SubGhzViewTransmitterModel * model,
@@ -237,6 +249,10 @@ bool subghz_view_transmitter_input(InputEvent* event, void* context) {
                 } else if(nice_flors_get_original_btn() != 0) {
                     furi_string_printf(
                         model->temp_button_id, "%01X", nice_flors_get_original_btn());
+                    model->draw_temp_button = true;
+                } else if(somfy_telis_get_original_btn() != 0) {
+                    furi_string_printf(
+                        model->temp_button_id, "%01X", somfy_telis_get_original_btn());
                     model->draw_temp_button = true;
                 }
             },